+++ UD09/src/ex06/Asiento.java
@@ -1,4 +1,4 @@
-package UD09;
+package src.ex06;
 
 public class Asiento {
